# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [auth]
      summary: List all admin users (super admin only)
      responses:
        "200":
          description: List of admin users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreateUser
      tags: [auth]
      summary: Create a new admin user (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: adminUpdateUser
      tags: [auth]
      summary: Update an admin user (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchAdminUserInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteUser
      tags: [auth]
      summary: Deactivate (soft-delete) an admin user (super admin only)
      responses:
        "204":
          description: Deactivated
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Cannot deactivate the last active super admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/courses:
    get:
      operationId: adminListCourses
      tags: [education]
      summary: List all courses (admin)
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreateCourse
      tags: [education]
      summary: Create a new course (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertCourse"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /course-registrations:
    post:
      operationId: createCourseRegistrationPublic
      tags: [education]
      summary: Submit a new course registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertCourseRegistration"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseRegistration"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/course-registrations:
    get:
      operationId: adminListCourseRegistrations
      tags: [education]
      summary: List all course registrations (admin)
      responses:
        "200":
          description: List of course registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseRegistration"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/donation-campaigns:
    get:
      operationId: adminListDonationCampaigns
      tags: [donations]
      summary: List all donation campaigns (admin)
      responses:
        "200":
          description: List of donation campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DonationCampaign"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreateDonationCampaign
      tags: [donations]
      summary: Create a new donation campaign (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertDonationCampaign"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationCampaign"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /donations/checkout:
    post:
      operationId: checkoutDonation
      tags: [donations]
      summary: Process a donation payment via Square
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutDonation"
      responses:
        "201":
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationTransaction"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Payment failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/donations/transactions:
    get:
      operationId: adminListDonationTransactions
      tags: [donations]
      summary: List donation transactions (admin)
      parameters:
        - name: campaignId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of donation transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DonationTransaction"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AdminRole:
      type: string
      enum:
        - super_admin
        - masjid_admin
        - education_admin
        - donation_admin
        - content_editor
        - read_only
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/AdminRole"
        phone:
          type: ["string", "null"]
        active:
          type: boolean
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - active
        - createdAt
...
    CreateAdminUserInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        role:
          $ref: "#/components/schemas/AdminRole"
        phone:
          type: ["string", "null"]
        active:
          type: boolean
      required:
        - email
        - password
        - name
        - role
    PatchAdminUserInput:
      type: object
      properties:
        name:
          type: string
        role:
          $ref: "#/components/schemas/AdminRole"
        phone:
          type: ["string", "null"]
        active:
          type: boolean
        password:
          type: string
          minLength: 8
...
    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        ageGroup:
          type: ["string", "null"]
        schedule:
          type: ["string", "null"]
        fee:
          type: ["string", "null"]
        capacity:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        published:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - published
        - createdAt
    InsertCourse:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ageGroup:
          type: ["string", "null"]
        schedule:
          type: ["string", "null"]
        fee:
          type: ["string", "null"]
        capacity:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        published:
          type: boolean
      required:
        - title
        - description
        - published
...
    CourseRegistration:
      type: object
      properties:
        id:
          type: string
        courseId:
          type: string
        studentName:
          type: string
        guardianName:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - courseId
        - studentName
        - email
        - status
        - createdAt
    InsertCourseRegistration:
      type: object
      properties:
        courseId:
          type: string
        studentName:
          type: string
        guardianName:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - courseId
        - studentName
        - email
...
    DonationCampaign:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        targetAmount:
          type: ["string", "null"]
        raisedAmount:
          type: string
        externalDonationUrl:
          type: ["string", "null"]
        active:
          type: boolean
        featured:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - raisedAmount
        - active
        - featured
        - createdAt
    InsertDonationCampaign:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        targetAmount:
          type: ["string", "null"]
        raisedAmount:
          type: string
        externalDonationUrl:
          type: ["string", "null"]
        active:
          type: boolean
        featured:
          type: boolean
      required:
        - title
        - description
        - raisedAmount
        - active
        - featured
...
    CheckoutDonation:
      type: object
      properties:
        campaignId:
          type: string
        amount:
          type: string
        sourceId:
          type: string
        donorName:
          type: ["string", "null"]
        donorEmail:
          type: ["string", "null"]
      required:
        - campaignId
        - amount
        - sourceId
    DonationTransaction:
      type: object
      properties:
        id:
          type: string
        campaignId:
          type: string
        amount:
          type: string
        donorName:
          type: ["string", "null"]
        donorEmail:
          type: ["string", "null"]
        squarePaymentId:
          type: string
        status:
          type: string
          enum: [succeeded, failed]
        createdAt:
          type: string
      required:
        - id
        - campaignId
        - amount
        - squarePaymentId
        - status
        - createdAt
...