# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRole } from './adminRole';

export interface AdminUser {
  id: string;
  email: string;
  name: string;
  role: AdminRole;
  /** @nullable */
  phone?: string | null;
  active: boolean;
  /** @nullable */
  lastLoginAt?: string | null;
  createdAt: string;
}