# lib/api-zod/src/generated/types/createAdminUserInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRole } from './adminRole';

export interface CreateAdminUserInput {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
  role: AdminRole;
  /** @nullable */
  phone?: string | null;
  active?: boolean;
}