# lib/api-zod/src/generated/types/patchAdminUserInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRole } from './adminRole';

export interface PatchAdminUserInput {
  name?: string;
  role?: AdminRole;
  /** @nullable */
  phone?: string | null;
  active?: boolean;
  /** @minLength 8 */
  password?: string;
}